requirements.txt
requests==2.31.0
python-dotenv==1.0.0
feedparser==6.0.10
pillow==10.1.0
moviepy==1.0.3
gtts==2.4.0
edge-tts>=6.1.0
elevenlabs>=1.0.0
importlib-metadata>=4.0  # Required for Python 3.9 compatibility with elevenlabs
pydub==0.25.1
ollama==0.1.7
newsapi-python==0.2.7
google-api-python-client==2.100.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
sentence-transformers>=2.2.0  # For semantic similarity in article grouping
scikit-learn>=1.0.0  # For cosine similarity calculation
google-generativeai>=0.8.0  # For Gemini API with Google Search grounding

